requirements.txt
wheel==0.40.0
matplotlib==3.7.1
scipy==1.10.1
numpy==1.24.3
Pillow==9.5.0
numpy-quaternion==2022.4.3
opencv-python==4.7.0.72
